Stop Trex gracefully
Process.Kill() kills the process immediately, without sending any signal to the process. It causes some problems on t-rex if GPU fine tuning parameters (like --lock-cclock) are used.

There is a Process.CloseMainWindow() to kill the process gracefully, but it requires a window (for that running process), which is not a case in TrexMinerGUI.

So, in order to kill an attached terminal process gracefully, TrexMinerGUI sends CTRL+C signal to the app.