Fix context menu on tray icon
The context menu on tray icon does not get updated every time, there is a bug on trayIcon.Click event that it does not gets fired always. Instead of binding the UpdateContextMenu method to trayIcon.Click, bind it to ContextMenu.Opening.

Also, fix the start/stop button while stopping the miner.